python: Configure PythonInit file for kit when VTK_ENABLE_KITS is ON
This commit introduces the file "vtkPythonKitInit.cxx.in" that
is configured to call the initialization of all module associated
with a given kit.

To verify that all kits were properly wrapped, the content
of the vtk module has been dumped to file and compared. Beside of
the name of the imported modules all object were identical.

In both case, the following has been done:
1) Start bin/vtkpython
2) Execute the following:
case = "vtk_with_kits" # of "vtk_without_kits"
contents = sorted(dir(vtk))
with open("/tmp/%s.txt" % case, 'w') as f:
    f.write("\n".join(contents))

Change-Id: I11000d1d6caf9329b208a4db45b762daeec9b531